@@ -1,10 +1,8 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const maxrects_packer_1 = require("./maxrects_packer");
-const Rectangle_1 = require("./geom/Rectangle");
-const abstract_bin_1 = require("./abstract_bin");
-class MaxRectsBin extends abstract_bin_1.Bin {
-    constructor(maxWidth = maxrects_packer_1.EDGE_MAX_VALUE, maxHeight = maxrects_packer_1.EDGE_MAX_VALUE, padding = 0, options = { smart: true, pot: true, square: true, allowRotation: false }) {
+import { EDGE_MAX_VALUE } from "./maxrects-packer";
+import { Rectangle } from "./geom/Rectangle";
+import { Bin } from "./abstract-bin";
+export class MaxRectsBin extends Bin {
+    constructor(maxWidth = EDGE_MAX_VALUE, maxHeight = EDGE_MAX_VALUE, padding = 0, options = { smart: true, pot: true, square: true, allowRotation: false }) {
         super();
         this.maxWidth = maxWidth;
         this.maxHeight = maxHeight;
@@ -15,10 +13,26 @@ class MaxRectsBin extends abstract_bin_1.Bin {
         this.verticalExpand = false;
         this.width = this.options.smart ? 0 : maxWidth;
         this.height = this.options.smart ? 0 : maxHeight;
-        this.freeRects.push(new Rectangle_1.Rectangle(0, 0, this.maxWidth + this.padding, this.maxHeight + this.padding));
-        this.stage = new Rectangle_1.Rectangle(0, 0, this.width, this.height);
+        this.freeRects.push(new Rectangle(this.maxWidth + this.padding, this.maxHeight + this.padding));
+        this.stage = new Rectangle(this.width, this.height);
     }
-    add(width, height, data) {
+    add(...args) {
+        let width;
+        let height;
+        let data;
+        let rect;
+        if (args.length === 1) {
+            if (typeof args[0] !== 'object')
+                throw new Error("MacrectsBin.add(): Wrong parameters");
+            rect = args[0];
+            width = rect.width;
+            height = rect.height;
+        }
+        else {
+            width = args[0];
+            height = args[1];
+            data = args.length > 2 ? args[2] : null;
+        }
         let node = this.findNode(width + this.padding, height + this.padding);
         if (node) {
             this.updateBinSize(node);
@@ -34,20 +48,27 @@ class MaxRectsBin extends abstract_bin_1.Bin {
             }
             this.pruneFreeList();
             this.verticalExpand = this.width > this.height ? true : false;
-            let rect = new Rectangle_1.Rectangle(node.x, node.y, width, height, node.rot);
-            rect.data = data;
+            if (!rect) {
+                rect = new Rectangle(width, height, node.x, node.y, node.rot);
+                rect.data = data;
+            }
+            else {
+                rect.x = node.x;
+                rect.y = node.y;
+                rect.rot = node.rot;
+            }
             this.rects.push(rect);
             return rect;
         }
         else if (!this.verticalExpand) {
-            if (this.updateBinSize(new Rectangle_1.Rectangle(this.width + this.padding, 0, width + this.padding, height + this.padding))
-                || this.updateBinSize(new Rectangle_1.Rectangle(0, this.height + this.padding, width + this.padding, height + this.padding))) {
-                return this.add(width, height, data);
+            if (this.updateBinSize(new Rectangle(width + this.padding, height + this.padding, this.width + this.padding, 0))
+                || this.updateBinSize(new Rectangle(width + this.padding, height + this.padding, 0, this.height + this.padding))) {
+                return rect ? this.add(rect) : this.add(width, height, data);
             }
         }
         else {
-            if (this.updateBinSize(new Rectangle_1.Rectangle(0, this.height + this.padding, width + this.padding, height + this.padding)) || this.updateBinSize(new Rectangle_1.Rectangle(this.width + this.padding, 0, width + this.padding, height + this.padding))) {
-                return this.add(width, height, data);
+            if (this.updateBinSize(new Rectangle(width + this.padding, height + this.padding, 0, this.height + this.padding)) || this.updateBinSize(new Rectangle(width + this.padding, height + this.padding, this.width + this.padding, 0))) {
+                return rect ? this.add(rect) : this.add(width, height, data);
             }
         }
         return undefined;
@@ -62,7 +83,7 @@ class MaxRectsBin extends abstract_bin_1.Bin {
             if (r.width >= width && r.height >= height) {
                 areaFit = r.width * r.height - width * height;
                 if (areaFit < score) {
-                    bestNode = new Rectangle_1.Rectangle(r.x, r.y, width, height);
+                    bestNode = new Rectangle(width, height, r.x, r.y);
                     score = areaFit;
                 }
             }
@@ -72,7 +93,7 @@ class MaxRectsBin extends abstract_bin_1.Bin {
             if (r.width >= height && r.height >= width) {
                 areaFit = r.width * r.height - height * width;
                 if (areaFit < score) {
-                    bestNode = new Rectangle_1.Rectangle(r.x, r.y, height, width, true); // Rotated node
+                    bestNode = new Rectangle(height, width, r.x, r.y, true); // Rotated node
                     score = areaFit;
                 }
             }
@@ -87,12 +108,12 @@ class MaxRectsBin extends abstract_bin_1.Bin {
         if (usedNode.x < freeRect.x + freeRect.width && usedNode.x + usedNode.width > freeRect.x) {
             // New node at the top side of the used node
             if (usedNode.y > freeRect.y && usedNode.y < freeRect.y + freeRect.height) {
-                let newNode = new Rectangle_1.Rectangle(freeRect.x, freeRect.y, freeRect.width, usedNode.y - freeRect.y);
+                let newNode = new Rectangle(freeRect.width, usedNode.y - freeRect.y, freeRect.x, freeRect.y);
                 this.freeRects.push(newNode);
             }
             // New node at the bottom side of the used node
             if (usedNode.y + usedNode.height < freeRect.y + freeRect.height) {
-                let newNode = new Rectangle_1.Rectangle(freeRect.x, usedNode.y + usedNode.height, freeRect.width, freeRect.y + freeRect.height - (usedNode.y + usedNode.height));
+                let newNode = new Rectangle(freeRect.width, freeRect.y + freeRect.height - (usedNode.y + usedNode.height), freeRect.x, usedNode.y + usedNode.height);
                 this.freeRects.push(newNode);
             }
         }
@@ -101,12 +122,12 @@ class MaxRectsBin extends abstract_bin_1.Bin {
             usedNode.y + usedNode.height > freeRect.y) {
             // New node at the left side of the used node.
             if (usedNode.x > freeRect.x && usedNode.x < freeRect.x + freeRect.width) {
-                let newNode = new Rectangle_1.Rectangle(freeRect.x, freeRect.y, usedNode.x - freeRect.x, freeRect.height);
+                let newNode = new Rectangle(usedNode.x - freeRect.x, freeRect.height, freeRect.x, freeRect.y);
                 this.freeRects.push(newNode);
             }
             // New node at the right side of the used node.
             if (usedNode.x + usedNode.width < freeRect.x + freeRect.width) {
-                let newNode = new Rectangle_1.Rectangle(usedNode.x + usedNode.width, freeRect.y, freeRect.x + freeRect.width - (usedNode.x + usedNode.width), freeRect.height);
+                let newNode = new Rectangle(freeRect.x + freeRect.width - (usedNode.x + usedNode.width), freeRect.height, usedNode.x + usedNode.width, freeRect.y);
                 this.freeRects.push(newNode);
             }
         }
@@ -169,8 +190,8 @@ class MaxRectsBin extends abstract_bin_1.Bin {
                 freeRect.height = height - freeRect.y;
             }
         }, this);
-        this.freeRects.push(new Rectangle_1.Rectangle(this.width + this.padding, 0, width - this.width - this.padding, height));
-        this.freeRects.push(new Rectangle_1.Rectangle(0, this.height + this.padding, width, height - this.height - this.padding));
+        this.freeRects.push(new Rectangle(width - this.width - this.padding, height, this.width + this.padding, 0));
+        this.freeRects.push(new Rectangle(width, height - this.height - this.padding, 0, this.height + this.padding));
         this.freeRects.forEach((freeRect, index) => {
             if (freeRect.width <= 0 || freeRect.height <= 0) {
                 this.freeRects.splice(index, 1);
@@ -179,5 +200,4 @@ class MaxRectsBin extends abstract_bin_1.Bin {
         this.pruneFreeList();
     }
 }
-exports.MaxRectsBin = MaxRectsBin;
-//# sourceMappingURL=maxrects_bin.js.map
+//# sourceMappingURL=maxrects-bin.js.map
