@@ -1,11 +1,9 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const Rectangle_1 = require("./geom/Rectangle");
-const maxrects_bin_1 = require("./maxrects_bin");
-const oversized_element_bin_1 = require("./oversized_element_bin");
-exports.EDGE_MAX_VALUE = 4096;
-exports.EDGE_MIN_VALUE = 128;
-class MaxRectsPacker {
+import { Rectangle } from "./geom/Rectangle";
+import { MaxRectsBin } from "./maxrects-bin";
+import { OversizedElementBin } from "./oversized-element-bin";
+export const EDGE_MAX_VALUE = 4096;
+export const EDGE_MIN_VALUE = 128;
+export class MaxRectsPacker {
     /**
      * Creates an instance of MaxRectsPacker.
      * @param {number} width of the output atlas (default is 4096)
@@ -14,30 +12,47 @@ class MaxRectsPacker {
      * @param {IOption} [options={}] (Optional) packing options
      * @memberof MaxRectsPacker
      */
-    constructor(width = exports.EDGE_MAX_VALUE, height = exports.EDGE_MAX_VALUE, padding = 0, options = { smart: true, pot: true, square: true, allowRotation: false }) {
+    constructor(width = EDGE_MAX_VALUE, height = EDGE_MAX_VALUE, padding = 0, options = { smart: true, pot: true, square: true, allowRotation: false }) {
         this.width = width;
         this.height = height;
         this.padding = padding;
         this.options = options;
         this.bins = [];
     }
-    /**
-     * Add a bin/rectangle object with data to packer
-     * @param {number} width of the input bin/rectangle
-     * @param {number} height of the input bin/rectangle
-     * @param {*} data custom data object
-     * @memberof MaxRectsPacker
-     */
-    add(width, height, data) {
-        if (width > this.width || height > this.height) {
-            this.bins.push(new oversized_element_bin_1.OversizedElementBin(width, height, data));
+    add(...args) {
+        let width;
+        let height;
+        let data;
+        if (args.length === 1) {
+            if (typeof args[0] !== 'object')
+                throw new Error("MacrectsPacker.add(): Wrong parameters");
+            const rect = args[0];
+            if (rect.width > this.width || rect.height > this.height) {
+                this.bins.push(new OversizedElementBin(rect));
+            }
+            else {
+                let added = this.bins.find(bin => bin.add(rect) !== undefined);
+                if (!added) {
+                    let bin = new MaxRectsBin(this.width, this.height, this.padding, this.options);
+                    bin.add(rect);
+                    this.bins.push(bin);
+                }
+            }
         }
         else {
-            let added = this.bins.find(bin => bin.add(width, height, data) !== undefined);
-            if (!added) {
-                let bin = new maxrects_bin_1.MaxRectsBin(this.width, this.height, this.padding, this.options);
-                bin.add(width, height, data);
-                this.bins.push(bin);
+            width = args[0];
+            height = args[1];
+            data = args.length > 2 ? args[2] : null;
+            if (width > this.width || height > this.height) {
+                this.bins.push(new OversizedElementBin(width, height, data));
+            }
+            else {
+                let added = this.bins.find(bin => bin.add(width, height, data) !== undefined);
+                if (!added) {
+                    let bin = new MaxRectsBin(this.width, this.height, this.padding, this.options);
+                    bin.add(width, height, data);
+                    this.bins.push(bin);
+                }
             }
         }
     }
@@ -48,7 +63,7 @@ class MaxRectsPacker {
      * @memberof MaxRectsPacker
      */
     addArray(rects) {
-        this.sort(rects).forEach(r => this.add(r.width, r.height, r.data));
+        this.sort(rects).forEach(rect => this.add(rect));
     }
     /**
      * Load bins to the packer, overwrite exist bins
@@ -58,13 +73,13 @@ class MaxRectsPacker {
     load(bins) {
         bins.forEach((bin, index) => {
             if (bin.maxWidth > this.width || bin.maxHeight > this.height) {
-                this.bins.push(new oversized_element_bin_1.OversizedElementBin(bin.width, bin.height, {}));
+                this.bins.push(new OversizedElementBin(bin.width, bin.height, {}));
             }
             else {
-                let newBin = new maxrects_bin_1.MaxRectsBin(this.width, this.height, this.padding, bin.options);
+                let newBin = new MaxRectsBin(this.width, this.height, this.padding, bin.options);
                 newBin.freeRects.splice(0);
                 bin.freeRects.forEach((r, i) => {
-                    newBin.freeRects.push(new Rectangle_1.Rectangle(r.x, r.y, r.width, r.height));
+                    newBin.freeRects.push(new Rectangle(r.width, r.height, r.x, r.y));
                 });
                 newBin.width = bin.width;
                 newBin.height = bin.height;
@@ -104,5 +119,4 @@ class MaxRectsPacker {
         return rects.slice().sort((a, b) => Math.max(b.width, b.height) - Math.max(a.width, a.height));
     }
 }
-exports.MaxRectsPacker = MaxRectsPacker;
-//# sourceMappingURL=maxrects_packer.js.map
+//# sourceMappingURL=maxrects-packer.js.map
